Emit valid schema when arrays in sample objects are empty.
Fixes a bug where an empty `items` array is specified in the schema
when the "training" objects all contain empty arrays. In other words,
the schema contains `{"type": "array", "items": []}`. The empty array
in the `items` entry is not valid.

The fix was to not emit the `items` entry if the array is empty.