Add `Enum` schema strategy
Add a new subclass of `SchemaStrategy` for supporting the 'enum`
keyword.

The `Enum` class keeps track of a `set` that contains all of the values
that were provided by the 'enum' keywords in the schemas. When the
builder is converted to a schema, this set is converted to a list.